pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dinocore"
version = "1.0.0"
authors = [
  { name="Alp Sakaci", email="erenalpsakaci@gmail.com" },
]
description = "dynamic config management tool"
readme = "README.md"
requires-python = ">=3.6"
classifiers = []
dependencies = [
    "PyYAML==6.0.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
